get rid of filter_parser.py. Move parser components into appropriate modules.
