+++ hera_cal/vis_clean.py
@@ -7,6 +7,7 @@ from collections import OrderedDict as odict
 import datetime
 from uvtools import dspec
 from copy import deepcopy
+import argparse
 
 
 from astropy import constants
@@ -1265,3 +1266,81 @@ def noise_eq_bandwidth(window, axis=-1):
             Noise equivalent bandwidth of the window
     """
     return np.sqrt(window.shape[axis] * np.max(window, axis=axis)**2 / np.sum(window**2, dtype=np.float, axis=axis))
+
+# ------------------------------------------
+# Here is an argparser with core arguments
+# needed for all types of xtalk and delay
+# filtering.
+# ------------------------------------------
+
+
+def _filter_argparser():
+    """
+    Core Arg parser for commandline operation of hera_cal.delay_filter and hera_cal.xtalk_filter
+    Returns
+    -------
+        Argparser with core (but not complete) functionality that is called by _linear_argparser and
+        _clean_argparser.
+    """
+    a = argparse.ArgumentParser(description="Perform delay filter of visibility data.")
+    a.add_argument("infilename", type=str, help="path to visibility data file to delay filter")
+    a.add_argument("--filetype_in", type=str, default='uvh5', help='filetype of input data files (default "uvh5")')
+    a.add_argument("--filetype_out", type=str, default='uvh5', help='filetype for output data files (default "uvh5")')
+    a.add_argument("--calfile", default=None, type=str, help="optional string path to calibration file to apply to data before delay filtering")
+    a.add_argument("--partial_load_Nbls", default=None, type=int, help="the number of baselines to load at once (default None means load full data")
+    a.add_argument("--res_outfilename", default=None, type=str, help="path for writing the filtered visibilities with flags")
+    a.add_argument("--clobber", default=False, action="store_true", help='overwrites existing file at outfile')
+    a.add_argument("--spw_range", type=int, default=None, nargs=2, help="spectral window of data to foreground filter.")
+    a.add_argument("--tol", type=float, default=1e-9, help='Threshold for foreground and xtalk subtraction (default 1e-9)')
+
+    return a
+
+
+# ------------------------------------------
+# Here are arg-parsers clean filtering
+# for both xtalk and foregrounds.
+# ------------------------------------------
+
+
+def _clean_argparser():
+    '''
+    Arg parser for commandline operation of hera_cal.delay_filter in various clean modes.
+
+    Returns
+    -------
+        Arg-parser for linear filtering. Still needs domain specific args (delay versus xtalk).
+    '''
+    a = _filter_argparser()
+    a.add_argument("--CLEAN_outfilename", default=None, type=str, help="path for writing the filtered model visibilities (with the same flags)")
+    a.add_argument("--filled_outfilename", default=None, type=str, help="path for writing the original data but with flags unflagged and replaced with filtered models wherever possible")
+    clean_options = a.add_argument_group(title='Options for CLEAN')
+    clean_options.add_argument("--window", type=str, default='blackman-harris', help='window function for frequency filtering (default "blackman-harris",\
+                              see uvtools.dspec.gen_window for options')
+    clean_options.add_argument("--skip_wgt", type=float, default=0.1, help='skips filtering and flags times with unflagged fraction ~< skip_wgt (default 0.1)')
+    clean_options.add_argument("--maxiter", type=int, default=100, help='maximum iterations for aipy.deconv.clean to converge (default 100)')
+    clean_options.add_argument("--edgecut_low", default=0, type=int, help="Number of channels to flag on lower band edge and exclude from window function.")
+    clean_options.add_argument("--edgecut_hi", default=0, type=int, help="Number of channels to flag on upper band edge and exclude from window function.")
+    clean_options.add_argument("--gain", type=float, default=0.1, help="Fraction of residual to use in each iteration.")
+    clean_options.add_argument("--alpha", type=float, default=.5, help="If window='tukey', use this alpha parameter (default .5).")
+    return a
+
+# ------------------------------------------
+# Here are are parsers for linear filters.
+# for xtalk and foregrounds.
+# ------------------------------------------
+
+
+def _linear_argparser():
+    '''
+    Arg parser for commandline operation of hera_cal.delay_filter in various linear modes.
+
+    Returns
+    -------
+        Arg-parser for linear filtering. Still needs domain specific args (delay versus xtalk)
+    '''
+    a = _filter_argparser()
+    cache_options = a.add_argument_group(title='Options for caching')
+    a.add_argument("--write_cache", default=False, action="store_true", help="if True, writes newly computed filter matrices to cache.")
+    a.add_argument("--cache_dir", type=str, default=None, help="directory to store cached filtering matrices in.")
+    a.add_argument("--read_cache", default=False, action="store_true", help="If true, read in cache files in directory specified by cache_dir.")
+    return a
