+++ scripts/dayenu_delay_filter_run.py
@@ -8,11 +8,8 @@
 
 from hera_cal import delay_filter
 import sys
-import pickle
-import random
-import filter_parser
 
-parser = filter_parser.delay_filter_argparser(mode='dayenu')
+parser = delay_filter.delay_filter_argparser(mode='dayenu')
 a = parser.parse_args()
 
 # set kwargs
